requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
qdrant-client==1.7.3
redis==5.0.1
psycopg2-binary==2.9.9
google-generativeai==0.3.2
beautifulsoup4==4.12.3
news-please
lxml==5.1.0